# frontend/public/openapi.json
{
  "components": {
    "schemas": {
...
      "AlchemyWebhookDoc": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "event": {
            "properties": {
              "activity": {
                "description": "Address Activity payload",
                "items": {
                  "properties": {
                    "blockNum": {
                      "description": "hex block number",
                      "type": "string"
                    },
                    "fromAddress": {
                      "type": "string"
                    },
                    "hash": {
                      "type": "string"
                    },
                    "rawContract": {
                      "properties": {
                        "address": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "address"
                      ],
                      "type": "object"
                    },
                    "toAddress": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "blockNum",
                    "fromAddress",
                    "hash",
                    "rawContract",
                    "toAddress"
                  ],
                  "type": "object"
                },
                "type": "array"
              },
              "data": {
                "anyOf": [
                  {
                    "properties": {
                      "block": {
                        "properties": {
                          "logs": {
                            "items": {
                              "properties": {
                                "account": {
                                  "properties": {
                                    "address": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "address"
                                  ],
                                  "type": "object"
                                },
                                "topics": {
                                  "items": {
                                    "type": "string"
                                  },
                                  "type": "array"
                                },
                                "transaction": {
                                  "properties": {
                                    "hash": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "hash"
                                  ],
                                  "type": "object"
                                }
                              },
                              "required": [
                                "account",
                                "topics",
                                "transaction"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "number": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "logs",
                          "number"
                        ],
                        "type": "object"
                      }
                    },
                    "required": [
                      "block"
                    ],
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Custom Webhook (GRAPHQL) payload"
              }
            },
            "type": "object"
          },
          "id": {
            "description": "the delivery id, e.g. whevt_...; a repeat within 24 hours is ignored",
            "type": "string"
          },
          "type": {
            "description": "GRAPHQL or ADDRESS_ACTIVITY",
            "enum": [
              "GRAPHQL",
              "ADDRESS_ACTIVITY"
            ],
            "type": "string"
          },
          "webhookId": {
            "description": "the webhook's id, e.g. wh_...",
            "type": "string"
          }
        },
        "required": [
          "event",
          "id",
          "type",
          "webhookId"
        ],
        "type": "object"
      },
...
      "Check": {
        "properties": {
          "ageSec": {
            "type": "integer"
          },
          "agrees": {
            "type": "boolean"
          },
          "comparable": {
            "type": "boolean"
          },
          "distanceM": {
            "type": "integer"
          },
          "loggerLatE6": {
            "type": "integer"
          },
          "loggerLonE6": {
            "type": "integer"
          }
        },
        "required": [
          "ageSec",
          "agrees",
          "comparable",
          "distanceM",
          "loggerLatE6",
          "loggerLonE6"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "stable machine-readable code, e.g. invalid_request, unauthorized, replayed, chain_rejected",
                "type": "string"
              },
              "message": {
                "description": "human-readable explanation; never contains internal detail",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Vessel": {
        "properties": {
          "crossCheck": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Check"
              },
              {
                "type": "null"
              }
            ]
          },
          "last": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VesselLast"
              },
              {
                "type": "null"
              }
            ]
          },
          "live": {
            "type": "boolean"
          },
          "mmsi": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "track": {
            "items": {
              "$ref": "#/components/schemas/VesselFix"
            },
            "type": "array"
          }
        },
        "required": [
          "crossCheck",
          "last",
          "live",
          "mmsi",
          "name",
          "track"
        ],
        "type": "object"
      },
      "VesselFix": {
        "properties": {
          "latE6": {
            "type": "integer"
          },
          "lonE6": {
            "type": "integer"
          },
          "timestamp": {
            "type": "integer"
          }
        },
        "required": [
          "latE6",
          "lonE6",
          "timestamp"
        ],
        "type": "object"
      },
      "VesselLast": {
        "properties": {
          "cogDegX10": {
            "type": "integer"
          },
          "latE6": {
            "type": "integer"
          },
          "lonE6": {
            "type": "integer"
          },
          "sogKnotsX10": {
            "type": "integer"
          },
          "timestamp": {
            "type": "integer"
          }
        },
        "required": [
          "cogDegX10",
          "latE6",
          "lonE6",
          "sogKnotsX10",
          "timestamp"
        ],
        "type": "object"
      },
      "VesselRequest": {
        "properties": {
          "issuedAt": {
            "type": "integer"
          },
          "mmsi": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "issuedAt",
          "mmsi",
          "signature"
        ],
        "type": "object"
      },
      "WebhookResult": {
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "block": {
            "description": "the highest block among them",
            "type": "integer"
          },
          "duplicate": {
            "description": "the delivery id was seen in the last 24 hours; nothing was done",
            "type": "boolean"
          },
          "matched": {
            "description": "logs or activities from CargoFlow contracts",
            "type": "integer"
          },
          "woken": {
            "description": "the indexer was asked to sync now (it re-reads the logs from the RPC; the payload itself is never applied)",
            "type": "boolean"
          }
        },
        "required": [
          "accepted",
          "matched",
          "woken"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminKey": {
        "description": "the operator's ADMIN_API_KEY",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "alchemySignature": {
        "description": "Alchemy Notify: hex HMAC-SHA256 of the raw request body with the webhook's signing key (ALCHEMY_WEBHOOK_SIGNING_KEYS)",
        "in": "header",
        "name": "X-Alchemy-Signature",
        "type": "apiKey"
      },
      "sourceSignature": {
        "description": "Signed by a registered evidence source. The signing string is\n\n```\nCARGOFLOW-V1\\n<METHOD>\\n<PATH>\\n<TIMESTAMP>\\n<hex sha256(body)>\n```\n\n`ed25519` sources sign it with Ed25519. `p256` sources sign it with ECDSA P-256 over SHA-256 (X-Signature is the DER or raw 64-byte r||s signature). `webauthn` sources run navigator.credentials.get with challenge = sha256(signing string) (32 raw bytes) and send the assertion: X-Signature = signature, X-WebAuthn-Authenticator-Data = authenticatorData, X-WebAuthn-Client-Data = clientDataJSON (all base64url); the backend checks type `webauthn.get`, the challenge, the registered RP id hash, the user-present flag and the signature over authenticatorData || sha256(clientDataJSON).",
        "in": "header",
        "name": "X-Signature",
        "type": "apiKey"
      },
      "walletSignature": {
        "description": "Not a header: the body carries `issuedAt` and `signature` (hex EIP-191 personal_sign) over the operation's `x-cargoflow-signed-message`. Declared as a scheme so clients can see which operations need a wallet.",
        "in": "header",
        "name": "X-Wallet-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CargoFlow turns signed shipment telemetry into on-chain financing decisions. Public reads need no credentials. Writes are authenticated in one of three ways: the operator's admin key, a party's wallet signature over a fixed message (see each operation's `x-cargoflow-signed-message`), or an evidence source's request signature (Ed25519, P-256 or WebAuthn). Errors are `{\"error\": {\"code\", \"message\"}}` with stable codes. Amounts are USDG base units (6 decimals) as decimal strings.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "CargoFlow API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/shipments/{id}/vessel": {
      "get": {
        "operationId": "getVessel",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vessel"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "The vessel, its AIS track and the logger cross-check",
        "tags": [
          "Shipments"
        ]
      },
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow vessel\nshipment: <id>\nmmsi: <mmsi>\nissued: <t>\n```",
        "operationId": "setVessel",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VesselRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "replaced the named vessel"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vessel"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Name the vessel (exporter)",
        "tags": [
          "Shipments"
        ],
        "x-cargoflow-signed-message": "CargoFlow vessel\nshipment: <id>\nmmsi: <mmsi>\nissued: <t>"
      }
    },
...
    "/v1/webhooks/alchemy": {
      "post": {
        "description": "Accepts Custom Webhook (GraphQL log filter) and Address Activity deliveries, at most 1 MB. The raw body must carry a valid `X-Alchemy-Signature` (hex HMAC-SHA256 with one of ALCHEMY_WEBHOOK_SIGNING_KEYS; 401 otherwise; 503 `webhook_unavailable` when none are set). A delivery id seen in the last 24 hours answers 200 with `duplicate: true` and does nothing. When a log or activity comes from a CargoFlow contract the indexer is woken to sync now (up to that block, once CONFIRMATIONS deep); the payload itself is never applied: the indexer re-reads the logs from the RPC, so a forged or replayed delivery cannot change state. Other addresses are ignored.",
        "operationId": "alchemyWebhook",
        "parameters": [
          {
            "description": "hex HMAC-SHA256 of the raw body with the webhook's signing key",
            "in": "header",
            "name": "X-Alchemy-Signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlchemyWebhookDoc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "alchemySignature": []
          }
        ],
        "summary": "Alchemy Notify deliveries: wake the indexer",
        "tags": [
          "System"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "this backend",
      "url": "/"
    }
  ]
}